# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/join": {
      "post": {
        "operationId": "AuthController_join",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입이 완료되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (이메일 중복 등)"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "summary": "회원가입",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "이메일 또는 비밀번호가 일치하지 않습니다."
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "summary": "사용자 로그인",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/social": {
      "post": {
        "operationId": "AuthController_socialRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "소셜 사용자 등록 또는 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "소셜 로그인 등록",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/duplicateEmail": {
      "post": {
        "operationId": "AuthController_duplicateEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/duplicateEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용 가능한 이메일입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/duplicateEmailDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 사용 중인 이메일입니다."
          }
        },
        "summary": "이메일 중복 검사",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/duplicateNickname": {
      "post": {
        "operationId": "AuthController_duplicateNickname",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/duplicateNicknameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용 가능한 닉네임입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/duplicateNicknameDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 사용 중인 닉네임입니다."
          }
        },
        "summary": "닉네임 중복 검사",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/email/verify/send": {
      "post": {
        "operationId": "AuthController_sendCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sendEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "인증 코드가 이메일로 전송되었습니다."
          },
          "400": {
            "description": "이미 인증 코드가 발송되어 있음 or 이메일 오류"
          }
        },
        "summary": "이메일 인증보내기",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/email/verify/confirm": {
      "post": {
        "operationId": "AuthController_confirmCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/verifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "이메일 인증 완료"
          },
          "400": {
            "description": "인증 코드가 없거나, 일치하지 않거나, 만료됨"
          }
        },
        "summary": "이메일 인증확인",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "로그아웃 되었습니다."
          },
          "401": {
            "description": "인증되지 않은 사용자입니다."
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "summary": "사용자 로그아웃",
        "tags": [
          "Auth"
        ]
      }
    },
    "/mentor/apply": {
      "post": {
        "operationId": "MentorsController_applyMentor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMentorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "멘토 신청 완료"
          },
          "500": {
            "description": "멘토 신청 중 오류가 발생했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 신청",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor/company": {
      "patch": {
        "operationId": "MentorsController_updateCompany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "회사명이 변경되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCompanyDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "회사명 변경",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor/expertise": {
      "patch": {
        "operationId": "MentorsController_updateExpertise",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpertiseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "전문 분야가 변경되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateExpertiseDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "전문 분야 변경",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor/position": {
      "patch": {
        "operationId": "MentorsController_updatePosition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "직책이 변경되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePositionDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "직책 변경",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor/career": {
      "patch": {
        "operationId": "MentorsController_updateCareer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCareerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "연차가 변경되었습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCareerDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 (형식 오류, 인증 실패 등)"
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          },
          "500": {
            "description": "서버 에러"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "연차 변경",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor/company-name": {
      "patch": {
        "operationId": "MentorsController_updatePublicCompanyName",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyHiddenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "공개 여부가 수정되었습니다."
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "회사명 공개 여부 설정",
        "tags": [
          "Mentor"
        ]
      }
    },
    "/mentor": {
      "get": {
        "operationId": "MentorsController_mentorProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "멘토 정보 입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorProfileResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "멘토 정보를 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 정보",
        "tags": [
          "Mentor"
        ]
      }
    },
...
    "/admin/mentors": {
      "get": {
        "operationId": "AdminController_getMentorList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멘토 신청 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MentorListDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "멘토 신청 정보를 찾을 수 없습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 신청 목록 조회",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/mentors/{id}": {
      "get": {
        "operationId": "AdminController_getMentorDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "멘토 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MentorDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "해당 멘토를 찾을 수 없습니다."
          },
          "500": {
            "description": "멘토 상세 정보를 불러오는 데 실패했습니다."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멘토 상세 정보 조회",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "커넥트 api문서",
    "description": "커넥트 개발을 위한 api문서 \n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json`,",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Bearer Token"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "이메일 (로그인용)"
          },
          "nickname": {
            "type": "string",
            "example": "john_doe",
            "description": "닉네임"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "사용자 이름"
          },
          "phone": {
            "type": "string",
            "example": "01012345678",
            "description": "휴대폰번호 (형식: 01012345678)"
          },
          "role": {
            "type": "string",
            "example": "mentee",
            "description": "사용자 등급 (MENTOR, MENTEE, ADMIN)",
            "enum": [
              "mentee",
              "mentor",
              "admin"
            ]
          },
          "image": {
            "type": "string",
            "example": "https://example.com/profile.jpg",
            "description": "프로필 이미지 URL"
          }
        },
        "required": [
          "email",
          "nickname",
          "name",
          "role"
        ]
      },
...
      "JoinDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "이메일 (로그인용)"
          },
          "nickname": {
            "type": "string",
            "example": "john_doe",
            "description": "닉네임"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "사용자 이름"
          },
          "phone": {
            "type": "string",
            "example": "01012345678",
            "description": "휴대폰번호 (형식: 01012345678)"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd!",
            "description": "비밀번호 (8자 이상, 영문자, 숫자, 특수문자 포함)"
          }
        },
        "required": [
          "email",
          "nickname",
          "name",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "이메일 (로그인용)"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd!",
            "description": "비밀번호 (8자 이상, 영문자, 숫자, 특수문자 포함)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SocialLoginDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "KAKAO",
            "description": "소셜 로그인 제공자 (KAKAO, NAVER, GOOGLE 등)",
            "enum": [
              "google",
              "kakao",
              "naver"
            ]
          },
          "socialId": {
            "type": "string",
            "example": "123456789",
            "description": "소셜 플랫폼에서의 사용자 고유 ID"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "사용자 이메일 주소"
          },
          "name": {
            "type": "string",
            "example": "홍길동",
            "description": "사용자 이름"
          },
          "image": {
            "type": "string",
            "example": "https://example.com/avatar.jpg",
            "description": "프로필 이미지 URL"
          }
        },
        "required": [
          "provider",
          "socialId",
          "email",
          "name"
        ]
      },
      "duplicateEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "이메일 (로그인용)"
          }
        },
        "required": [
          "email"
        ]
      },
      "duplicateNicknameDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "john_doe",
            "description": "닉네임"
          }
        },
        "required": [
          "nickname"
        ]
      },
      "sendEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "인증용 이메일"
          }
        },
        "required": [
          "email"
        ]
      },
      "verifyCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "인증용 이메일"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "이메일로 전송된 인증 코드"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "CreateMentorDto": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string",
            "example": "Google",
            "description": "소속 회사명"
          },
          "introduce": {
            "type": "string",
            "example": "10년차 프론트엔드 개발자입니다.",
            "description": "멘토 자기소개"
          },
          "position": {
            "type": "string",
            "example": "backend",
            "description": "직책",
            "enum": [
              "backend",
              "frontend",
              "fullstack",
              "mobile",
              "devops",
              "security",
              "data_engineer",
              "data_scientist",
              "pm_po",
              "ux_designer",
              "ui_designer",
              "marketing",
              "hr",
              "cs",
              "finance",
              "etc"
            ]
          },
          "expertise": {
            "type": "array",
            "example": [
              "consulting",
              "design"
            ],
            "description": "전문 분야 (다중 선택)",
            "items": {
              "type": "string",
              "enum": [
                "business",
                "marketing",
                "sales",
                "finance",
                "hr",
                "it",
                "design",
                "consulting",
                "manufacturing",
                "etc"
              ]
            }
          },
          "career": {
            "type": "string",
            "example": "junior",
            "description": "연차",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ]
          },
          "portfolio": {
            "type": "string",
            "example": "https://portfolio.example.com",
            "description": "포트폴리오 페이지 url"
          }
        },
        "required": [
          "introduce",
          "position",
          "expertise",
          "career",
          "portfolio"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string",
            "example": "Google",
            "description": "소속 회사명"
          }
        }
      },
      "UpdateExpertiseDto": {
        "type": "object",
        "properties": {
          "expertise": {
            "type": "array",
            "example": [
              "consulting",
              "design"
            ],
            "description": "전문 분야 (다중 선택)",
            "items": {
              "type": "string",
              "enum": [
                "business",
                "marketing",
                "sales",
                "finance",
                "hr",
                "it",
                "design",
                "consulting",
                "manufacturing",
                "etc"
              ]
            }
          }
        },
        "required": [
          "expertise"
        ]
      },
      "UpdatePositionDto": {
        "type": "object",
        "properties": {
          "position": {
            "type": "string",
            "example": "backend",
            "description": "직책",
            "enum": [
              "backend",
              "frontend",
              "fullstack",
              "mobile",
              "devops",
              "security",
              "data_engineer",
              "data_scientist",
              "pm_po",
              "ux_designer",
              "ui_designer",
              "marketing",
              "hr",
              "cs",
              "finance",
              "etc"
            ]
          }
        },
        "required": [
          "position"
        ]
      },
      "UpdateCareerDto": {
        "type": "object",
        "properties": {
          "career": {
            "type": "string",
            "example": "junior",
            "description": "연차",
            "enum": [
              "junior",
              "middle",
              "senior",
              "lead"
            ]
          }
        },
        "required": [
          "career"
        ]
      },
      "UpdateCompanyHiddenDto": {
        "type": "object",
        "properties": {
          "isCompanyHidden": {
            "type": "boolean",
            "example": true,
            "description": "회사명 공개 여부"
          }
        },
        "required": [
          "isCompanyHidden"
        ]
      },
      "MentorProfileResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "멘토 정보 입니다.",
            "description": "응답 메시지"
          },
          "career": {
            "type": "string",
            "example": "5년차 백엔드 개발자",
            "description": "경력"
          },
          "company": {
            "type": "string",
            "example": "네이버",
            "description": "회사명"
          },
          "expertise": {
            "type": "string",
            "example": "백엔드 개발, 시스템 설계",
            "description": "전문 분야"
          },
          "isCompanyHidden": {
            "type": "boolean",
            "example": false,
            "description": "회사명 공개 여부 (true = 숨김)"
          },
          "position": {
            "type": "string",
            "example": "시니어 엔지니어",
            "description": "직무/직책"
          }
        },
        "required": [
          "message",
          "career",
          "company",
          "expertise",
          "isCompanyHidden",
          "position"
        ]
      },
...
      "MentorListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b1a8f3e1-3b0f-4e9b-98a2-c4f0e6d3a3b4",
            "description": "사용자 UUID"
          },
          "expertise": {
            "type": "string",
            "example": "React, Next.js",
            "description": "전문 분야"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "이메일 (로그인용)"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "사용자 이름"
          },
          "status": {
            "type": "string",
            "example": "pending",
            "description": "승인 여부",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-05-09T12:34:56Z",
            "description": "멘토 등록일"
          }
        },
        "required": [
          "id",
          "expertise",
          "email",
          "name",
          "status"
        ]
      },
...
    }
  }
}